+++ runtime/digi00x/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "digi00x-runtime"
+name = "snd-firewire-digi00x-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementation of service runtime for models in Digidesign Digi 00x Family
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA firewire-digi00x
+driver (snd-firewire-digi00x) supports.
 """
 license = "GPL-3.0-or-later"
 publish = false
