+++ runtime/digi00x/src/main.rs
@@ -4,7 +4,8 @@ mod model;
 
 use {
     alsactl::{prelude::*, *},
-    core::{card_cntr::*, dispatcher::*, *},
+    clap::Parser,
+    core::{card_cntr::*, cmdline::*, dispatcher::*, LogLevel, *},
     firewire_digi00x_protocols as protocols,
     glib::{
         source, {Error, FileError},
@@ -36,7 +37,7 @@ enum Model {
     Digi003(Digi003Model),
 }
 
-pub struct Dg00xRuntime {
+struct Dg00xRuntime {
     unit: (SndDigi00x, FwNode),
     model: Model,
     card_cntr: CardCntr,
@@ -342,4 +343,27 @@ impl Dg00xRuntime {
             self.timer = None;
         }
     }
+}
+
+struct Dg00xServiceCmd;
+
+#[derive(Parser, Default)]
+#[clap(name = "snd-firewire-digi00x-ctl-service")]
+struct Arguments {
+    /// The numeric identifier of sound card in Linux sound subsystem.
+    card_id: u32,
+
+    /// The level to debug runtime, disabled as a default.
+    #[clap(long, short, arg_enum)]
+    log_level: Option<LogLevel>,
+}
+
+impl ServiceCmd<Arguments, u32, Dg00xRuntime> for Dg00xServiceCmd {
+    fn params(args: &Arguments) -> (u32, Option<LogLevel>) {
+        (args.card_id, args.log_level)
+    }
+}
+
+fn main() {
+    Dg00xServiceCmd::run()
 }
